requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pinecone-client==2.2.4  # Update back to newer version
typing-extensions==4.7.1  # Add explicit version
langchain==0.0.350
sentence-transformers==2.2.2
openai==1.3.5
pydantic==2.5.2
gunicorn==21.2.0
